[Snippets] Applied Vladislav comments
