+++ src/common/snippets/include/snippets/lowered/pass/insert_specific_iterations.hpp
@@ -10,6 +10,7 @@
 #include "snippets/lowered/loop_manager.hpp"
 #include "snippets/op/loop.hpp"
 
+
 namespace ov {
 namespace snippets {
 namespace lowered {
