+++ src/plugins/intel_cpu/src/transformations/snippets/x64/pass/lowered/brgemm_cpu_blocking.hpp
@@ -20,6 +20,22 @@ class BrgemmCPUBlocking : public ov::snippets::lowered::pass::BrgemmBlocking<Brg
 public:
     OPENVINO_RTTI("BrgemmCPUBlocking", "BrgemmBlocking")
 
+    /**
+     * @interface DummyPass
+     * @brief The empty pass which is used to force insertion of first specific iteration of loop by K dimension
+     *        Note: the pass is in public section to have opportunity to validate blocking loop in tests
+     * @ingroup snippets
+     */
+    class DummyPass : public snippets::lowered::pass::RangedPass {
+    public:
+        DummyPass() = default;
+        OPENVINO_RTTI("DummyPass", "RangedPass")
+        bool run(snippets::lowered::LinearIR& linear_ir,
+                 snippets::lowered::LinearIR::constExprIt begin,
+                 snippets::lowered::LinearIR::constExprIt end) override;
+        std::shared_ptr<snippets::lowered::pass::PassBase> merge(const std::shared_ptr<snippets::lowered::pass::PassBase>& other) override;
+    };
+
 private:
     static snippets::lowered::LinearIR::constExprIt move_new_memory_buffer(snippets::lowered::LinearIR& linear_ir,
                                                                            const snippets::lowered::LinearIR::constExprIt& brgemm_it);
