+++ src/plugins/intel_cpu/src/transformations/tpp/x64/pass/lowered/brgemm_tpp_blocking.hpp
@@ -21,10 +21,12 @@ class BrgemmTPPBlocking : public ov::snippets::lowered::pass::BrgemmBlocking<ov:
 public:
     OPENVINO_RTTI("BrgemmTPPBlocking", "BrgemmBlockingBase")
 
-private:
-    std::tuple<size_t, size_t, size_t> get_blocking_params(const ov::snippets::lowered::ExpressionPtr& brgemm_expr) override;
-    ov::snippets::lowered::SpecificIterationHandlers get_k_loop_handlers(size_t work_amount, size_t block_size) const override;
-
+    /**
+     * @interface SetBrgemmBeta
+     * @brief The pass set `beta = 0` to BrgemmTPP.
+     *        Note: the pass is in public section to have opportunity to validate blocking loop in tests
+     * @ingroup snippets
+     */
     class SetBrgemmBeta : public snippets::lowered::pass::RangedPass {
     public:
         SetBrgemmBeta() = default;
@@ -34,6 +36,10 @@ private:
                  ov::snippets::lowered::LinearIR::constExprIt end) override;
         std::shared_ptr<snippets::lowered::pass::PassBase> merge(const std::shared_ptr<snippets::lowered::pass::PassBase>& other) override;
     };
+
+private:
+    std::tuple<size_t, size_t, size_t> get_blocking_params(const ov::snippets::lowered::ExpressionPtr& brgemm_expr) override;
+    ov::snippets::lowered::SpecificIterationHandlers get_k_loop_handlers(size_t work_amount, size_t block_size) const override;
 };
 
 }  // namespace pass
